+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Demo coverage report endpoint in new `report` module
 - Coverage completeness lines in HTML coverage report
 - Default threshold level coverage lines in HTML coverage report
+- Hidden table cell showing incompletely covered genes in coverage report
 ### Changed
 - Moved helper function from endpoints coverage to crud samples
 - Deleted unused `src/chanjo2/meta/handle_query_intervals.py` file